Fix libff Mac OS build
libff was being installed with a "libff.dylib" target install name.
This caused trouble when linking other libraries, because it made it
impossible to find. This commit passes CMAKE_INSTALL_NAME_DIR to
make it have an absolute path, similarly to how the secp256k1 library
is built.

This also changes cache tags to invalidate them and get rid of broken
binaries which would otherwise not be rebuilt.