GitHub Actions: improve macOS releases (#527)
* GitHub Actions: invalidate caches

* GitHub Actions: workaround for actions/virtual-environments#1811

Currently the CI system fails on macOS with several errors related
to Homebrew. This is a temporal workaround for the issue. See

    https://github.com/actions/virtual-environments/issues/1811

for more details.

* GitHub Actions: workaround for actions/cache#403

Currently the CI system fails due to cache corruption on macOS.
This can be temporarily worked around by using GNU tar instead
of BSD tar. See

    https://github.com/actions/cache/issues/403

for more details.

* Re-enable macOS CI

* GitHub Actions: improve macOS releases

Currently, releases for macOS are not very portable. The
binaries have hardcoded paths and might not run on other
systems. This commit improves the macOS binary releases
made via GitHub actions by bundling binary copies of libff,
libgmp and libsecp256k1, and by adjusting the rpath and
library names to make echidna run on other systems.