Move echidna action to crytic/echidna-action
This removes the action from the echidna repository. The action will
now live on its own repository, where it can be versioned and published
independently.

Additionally, this updates the workflow to refer to the action on its
new location, as well as adds a few lines to the README pointing to the
new action location.