ci: remove `mstksg/get-package` (#952)
* ci: remove `mstksg/get-package`

This action has not been updated since 2019, and is currently causing
some deprecation warnings to show up. This replaces it with a few run
steps for each OS.

* ci: drop redundant Linux dependencies

These are already pre-installed on the GitHub runners:

    autoconf is already the newest version (2.69-11.1).
    automake is already the newest version (1:1.16.1-4ubuntu6).
    libtool is already the newest version (2.4.6-14).

* ci: adjust caching

The stack rule missed the local workdir, and some keys were missing
hashes to make them get cleared when necessary.