Add labels to call sequence
hevm already supports the `vm.label(address, string)` cheatcode, but
Echidna has not made use of the labels so far in its output. This adds
some basic support when printing call sequences. When available, it
will show the corresponding label next to any from/to address, or
next to any address arguments.