Fix and reorganize shrinking
The current shrinking implementation had a bug/limitation: it always
tries to shrink the sender. In other words, if a transaction has a
sender that is not minimal, Echidna will try to reduce it. However,
the transaction might require a different sender to still cause
the assertion failure, and the fact that Echidna will unconditionally
try to lower it might cause the shrinking to get 'stuck'.

This commit reworks the shrinking process to try and favor a more
incremental shrinking, with multiple passes of smaller changes. The
old process consisted of:

 - 50% chance of removing one tx
 - 50% chance of (reducing sender _and_ shrinking) _all_ txs

The new process consists of, for each transaction independently:

 - 35% chance of shrinking the transaction
 - 30% chance of removing the transaction
 - 30% chance of keeping the transaction as-is
 -  5% chance of reducing sender

"wait" calls are treated differently:

 - 60% chance of reducing or removing delay
 - 40% chance of keeping it as-is

Fixes: #1224